PP4.13ag5: src/gauge/GaugeEligibility.sol - numeratorSma and denominatorSma narrowed to uint128 per PP-D56 (xi), adjacent so solc packs them into one slot and RankedEntry drops from four slots to three; efficiencyRatio stays uint256 as the sort key with no bound argument behind it, and the event ABI is untouched since uint128 widens implicitly for the emit

+++ src/gauge/GaugeEligibility.sol
@@ -100,8 +100,8 @@ contract GaugeEligibility is IGaugeEligibility {
     ///      without re-reading the oracle. Storage, because accumulation spans transactions.
     struct RankedEntry {
         address pool;
-        uint256 numeratorSma;
-        uint256 denominatorSma;
+        uint128 numeratorSma;
+        uint128 denominatorSma;
         uint256 efficiencyRatio;
     }
